+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Changelog
   This is to fix the ``/admin/xds_dump`` endpoint from requiring auth.
 * Added a bit more detail (what XDS type, what envoy version) for errors produced when a template fails
   to be parsed as valid YAML.
+* Bugfix ``/admin/xds_dump`` view not awaiting discovery response coroutine
+* Updated mock_discovery request to return the DiscoveryRequest dataclass instead of dictionary
 
 0.2.2 2019-08-23
 ----------------